+++ app/config/colors.js
@@ -1,0 +1,24 @@
+export default {
+    primary: '#ab47bc',
+    primaryLight: '#df78ef',
+    primaryDark: '#790e8b',
+    secondary: '#d32f2f',
+    secondaryLight: '#ff6659',
+    secondaryDark: '#9a0007',
+
+    /*Dark text on light background*/
+    primaryText: 'rgba(0, 0, 0, 0.87)',
+    secondaryText: 'rgba(0, 0, 0, 0.54)',
+    hintText: 'rgba(0, 0, 0, 0.38)',
+
+    background: '#FAFAFA',
+    highBackground: '#FFFFFF',
+
+    /*Light text on dark background*/
+    primaryTextDark: 'rgba(255, 255, 255, 1)',
+    secondaryTextDark: 'rgba(255, 255, 255, 0.7)',
+    hintTextDark: 'rgba(255, 255, 255, 0.5)',
+
+    backgroundDark: '#303030',
+    highBackgroundDark: '#424242',
+}
