+++ app/config/styles.js
@@ -1,12 +1,14 @@
 import { StyleSheet } from 'react-native';
+import colors from './colors';
 
 export default StyleSheet.create({
     mainContainer: {
         flex: 1,
         marginTop: 56,
+        backgroundColor: colors.background,
     },
     navBar: {
-        backgroundColor: '#9c27b0',
+        backgroundColor: colors.primary,
         flex: 1,
         height: 56,
         paddingHorizontal: 16,
@@ -14,4 +16,7 @@ export default StyleSheet.create({
         alignItems: 'center',
         elevation: 4,
     },
+    text: {
+        color: colors.primaryText,
+    }
 });
